+++ resources/views/bootstrap-components/form/toggle.blade.php
@@ -6,7 +6,7 @@
            type="checkbox"
            name="{{ $name }}"
         {{ htmlAttributes($componentHtmlAttributes) }}>
-    <label class="custom-control-label" for="{{ $componentId }}">{{ $label }}</label>
+    <label class="custom-control-label" for="{{ $componentId }}">@if($icon)<span class="{{ $componentId }}-icon">{!! $icon !!}</span> @endif{{ $label }}</label>
     @include('bootstrap-components::bootstrap-components.partials.validation-feedback')
     @include('bootstrap-components::bootstrap-components.partials.legend')
 </div>