+++ tests/Unit/Form/CheckboxTest.php
@@ -17,6 +17,7 @@ class CheckboxTest extends BootstrapComponentsTestCase
         $this->assertTrue(array_key_exists('checkbox', config('bootstrap-components.form')));
         // components.form.checkbox
         $this->assertTrue(array_key_exists('view', config('bootstrap-components.form.checkbox')));
+        $this->assertTrue(array_key_exists('icon', config('bootstrap-components.form.checkbox')));
         $this->assertTrue(array_key_exists('legend', config('bootstrap-components.form.checkbox')));
         $this->assertTrue(array_key_exists('class', config('bootstrap-components.form.checkbox')));
         $this->assertTrue(array_key_exists('html_attributes', config('bootstrap-components.form.checkbox')));
@@ -60,6 +61,43 @@ class CheckboxTest extends BootstrapComponentsTestCase
         $user = $this->createUniqueUser();
         $html = bsCheckbox()->model($user)->name('name')->toHtml();
         $this->assertContains('checked="checked"', $html);
+    }
+
+    public function testConfigIcon()
+    {
+        $configIcon = 'test-config-icon';
+        config()->set('bootstrap-components.form.checkbox.icon', $configIcon);
+        $html = bsCheckbox()->name('name')->toHtml();
+        $this->assertContains(
+            '<label class="custom-control-label" for="checkbox-name"><span class="checkbox-name-icon">'
+            . $configIcon . '</span> validation.attributes.name</label>',
+            $html
+        );
+    }
+
+    public function testSetIcon()
+    {
+        $configIcon = 'test-config-icon';
+        $customIcon = 'test-custom-icon';
+        config()->set('bootstrap-components.form.checkbox.icon', $configIcon);
+        $html = bsCheckbox()->name('name')->icon($customIcon)->toHtml();
+        $this->assertNotContains(
+            '<label class="custom-control-label" for="checkbox-name"><span class="checkbox-name-icon">'
+            . $configIcon . '</span> validation.attributes.name</label>',
+            $html
+        );
+        $this->assertContains(
+            '<label class="custom-control-label" for="checkbox-name"><span class="checkbox-name-icon">'
+            . $customIcon . '</span> validation.attributes.name</label>',
+            $html
+        );
+    }
+
+    public function testNoIcon()
+    {
+        config()->set('bootstrap-components.form.checkbox.icon', null);
+        $html = bsCheckbox()->name('name')->toHtml();
+        $this->assertNotContains('<span class="checkbox-name-icon">', $html);
     }
 
     public function testChecked()